pyproject.toml
###########
# 📜 Poetry
###########
[tool.poetry]
name = "ciot"
version = "0.1.0"
description = "Python demos for CIOT Diploma in Tax Technology."
authors = ["Coefficient Systems Ltd <contact@coefficient.ai>"]
license = "UNLICENSED"
classifiers = ["Private :: Do Not Upload"]
packages = [
  { include = "ciot", from = "ciot" },
  # Add more packages here, e.g.
  # { include = "module_name", from = "ciot" },
]

[tool.poetry.dependencies]
python = "^3.10.9"
google-cloud-bigquery = "^3.4.1"
google-cloud-functions = "^1.9.0"
# Everything below here is alphabetically sorted

[tool.poetry.dev-dependencies]
# Everything below here is alphabetically sorted
black = "^22.8.0"
detect-secrets = "1.2.0"
flake8 = "^4.0.1"
flake8-docstrings = "^1.6.0"
flake8-eradicate = "^1.3.0"
flake8-fixme = "^1.1.1"
flake8-implicit-str-concat = "^0.3.0"
flake8-no-pep420 = "^2.3.0"
flake8-print = "^5.0.0"
flake8-return = "^1.1.3"
ipdb = "^0.13.9"
isort = "^5.10.1"
jupyter-black = "^0.3.1"
pip-audit = "^2.4.4"
poetryup = "^0.8.2"
pre-commit = "^2.20.0"
pylint = "^2.15.2"
pytest = "^7.1.3"
towncrier = "^21.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

############
# ✅ Linters
############
[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
default_section = "THIRDPARTY"
known_first_party = ["ciot"]
sections = "FUTURE,STDLIB,THIRDPARTY,FIRSTPARTY,LOCALFOLDER"
add_imports = "from __future__ import annotations"

[tool.pylint.MASTER]
ignore-paths = [".git"]
load-plugins = []

[tool.pylint."MESSAGES CONTROL"]
enable = "all"
max-module-lines = 2000
max-line-length = 100
max-locals = 50
min-similarity-lines = 150
max-statements = 89
max-args = 22
max-branches = 17
# good-names = []
disable = ['broad-except', 'locally-disabled', 'suppressed-message']
logging-format-style = "new"

##############
# 📣 Towncrier
##############
[tool.towncrier]
package = "ciot"
package_dir = "."
filename = "CHANGELOG.md"
directory = ".changelog/"
template = ".changelog/template.md"
title_format = "## [{version}] - {project_date}"
issue_format = "[#{issue}](https://github.com/CoefficientSystems/ciot-python-demo/issues/{issue})"
start_string = "<!-- TOWNCRIER -->\n"
underlines = ["", ""]

# .added for new features
# .changed for changes in existing functionality
# .deprecated for soon-to-be removed features
# .removed for now removed features
# .fixed for any bug fixes
# .security in case of vulnerabilities
# .analysis for data analyses
# .docs for documentation improvements
# .maintenance for maintenance tasks & upgrades
[[tool.towncrier.type]]
directory = "added"
name = "Added"
showcontent = true

[[tool.towncrier.type]]
directory = "changed"
name = "Changed"
showcontent = true

[[tool.towncrier.type]]
directory = "deprecated"
name = "Deprecated"
showcontent = true

[[tool.towncrier.type]]
directory = "removed"
name = "Removed"
showcontent = true

[[tool.towncrier.type]]
directory = "fixed"
name = "Bug Fixes"
showcontent = true

[[tool.towncrier.type]]
directory = "security"
name = "Security"
showcontent = true

[[tool.towncrier.type]]
directory = "analysis"
name = "Analyses"
showcontent = true

[[tool.towncrier.type]]
directory = "docs"
name = "Improved Documentation"
showcontent = true

[[tool.towncrier.type]]
directory = "maintenance"
name = "Maintenance Changes"
showcontent = true
